+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ Version 1.4.0 (Lupin): Not yet released
 - Fix incorrect clickable area size for launcher FFLink button
 
 [@is-this-c](https://github.com/is-this-c)
+- Disable weapon cycle selection, if `Mouse 3` is pressed
 - For `Run` games, rename `Score` column to `Deaths`, and compare `Loads` in `std::ranges::sort`
 
 Version 1.3.0 (Bakeapple): Released Apr-22-2026